change market_data.go to products.go and move tests to products_test.go
