+++ pkg/godax/products.go
